Sun, Jan 23, 2022, 3:05 AM +00:00

+++ CS50x/week1/pset1/credit/tester.c
@@ -2,18 +2,18 @@
 #include <cs50.h>
 #include <math.h>
 
-int getLengthOfLong();
+double getLengthOfLong();
 
 int main(void) {
 	double test = get_long("try me, user: ");
-	int len = getLengthOfLong(test);
+	double len = getLengthOfLong(test);
     // long len = test%10;
-	printf("%i\n", len);
+	printf("%f\n", len);
     return 1;
 }
 
-int getLengthOfLong(loong) {
-	int len = log10(loong);
+double getLengthOfLong(loong) {
+	double len = log10(loong);
 
     return len;
 }
